+++ routes/college.js
@@ -8,23 +8,20 @@ const isEmpty = require("lodash/isEmpty");
 // @desc    Get All College
 // @access  Public
 router.get("/", async (req, res) => {
-  const condition = !isNil(req.query.condition) ? JSON.parse(req.query.condition) : {};
+  const condition = !isNil(req.query.condition)
+    ? JSON.parse(req.query.condition)
+    : {};
   if (isNil(condition.deletedAt)) {
-      condition.deletedAt = {
-          $exists: false
-      }
+    condition.deletedAt = {
+      $exists: false,
+    };
   }
   try {
     const getAllCollege = await College.find(condition);
-    res.json({
-      dbRes: getAllCollege,
-      isSuccess: true
-    });
-  } catch (error) {
-    res.json({
-      dbRes: error.message,
-      isSuccess: false
-    });
+    res.json(getAllCollege);
+  } catch ({ message: errMessage }) {
+    const message = errMessage ? errMessage : UNKNOWN_ERROR_OCCURED;
+    res.status(500).json(message);
   }
 });
 
@@ -36,17 +33,17 @@ router.get("/:id", async (req, res) => {
     const getCollege = await College.findById({
       _id: req.params.id,
       deletedAt: {
-        $exists: false
-      }
+        $exists: false,
+      },
     });
     res.json({
       dbRes: getCollege,
-      isSuccess: true
+      isSuccess: true,
     });
   } catch (error) {
     res.json({
       dbRes: error.message,
-      isSuccess: false
+      isSuccess: false,
     });
   }
 });
@@ -60,38 +57,27 @@ router.post("/", async (req, res) => {
   if (!isNil(collegeName) && !isNil(collegeCode)) {
     const newCollege = new College({
       collegeName,
-      collegeCode
+      collegeCode,
     });
     try {
       const getCollege = await College.find({
         collegeName,
         deletedAt: {
-          $exists: false
-        }
+          $exists: false,
+        },
       });
       if (getCollege.length === 0) {
         const createCollege = await newCollege.save();
-        res.json({
-          dbRes: createCollege,
-          isSuccess: true
-        });
+        res.json(createCollege);
       } else {
-        res.json({
-          dbRes: "College name must be unique",
-          isSuccess: false
-        });
+        res.json(res.status(500).json("College is already in use"));
       }
-    } catch (error) {
-      res.json({
-        dbRes: error.message,
-        isSuccess: false
-      });
+    } catch ({ message: errMessage }) {
+      const message = errMessage ? errMessage : UNKNOWN_ERROR_OCCURED;
+      res.status(500).json(message);
     }
   } else {
-    res.json({
-      dbRes: "Required values are either invalid or empty",
-      isSuccess: false
-    });
+    res.status(500).json("Required values are either invalid or empty");
   }
 });
 
@@ -99,15 +85,15 @@ router.post("/", async (req, res) => {
 // @desc    Update A College
 // @access  Private
 router.put("/:id", async (req, res) => {
-    const collegeName = req.body.collegeName;
-    const collegeCode = req.body.collegeCode;
+  const collegeName = req.body.collegeName;
+  const collegeCode = req.body.collegeCode;
   if (!isNil(collegeName) && !isNil(collegeCode)) {
     try {
       const getCollege = await College.find({
         collegeName,
         deletedAt: {
-          $exists: false
-        }
+          $exists: false,
+        },
       });
       if (getCollege.length === 0) {
         const updateCollege = await College.findByIdAndUpdate(req.params.id, {
@@ -119,24 +105,24 @@ router.put("/:id", async (req, res) => {
         });
         res.json({
           dbRes: updateCollege,
-          isSuccess: true
+          isSuccess: true,
         });
       } else {
         res.json({
           dbRes: "College name must be unique",
-          isSuccess: false
+          isSuccess: false,
         });
       }
     } catch (error) {
       res.json({
         dbRes: error.message,
-        isSuccess: false
+        isSuccess: false,
       });
     }
   } else {
     res.json({
       dbRes: "Required values are either invalid or empty",
-      isSuccess: false
+      isSuccess: false,
     });
   }
 });
@@ -148,25 +134,17 @@ router.patch("/:id", async (req, res) => {
   const condition = req.body;
   if (!isEmpty(condition)) {
     try {
-        const updateCollege = await College.findByIdAndUpdate(req.params.id, {
-          $set: condition,
-          updatedAt: Date.now(),
-        });
-        res.json({
-          dbRes: updateCollege,
-          isSuccess: true
-        });
-    } catch (error) {
-      res.json({
-        dbRes: error.message,
-        isSuccess: false
+      const updateCollege = await College.findByIdAndUpdate(req.params.id, {
+        $set: condition,
+        updatedAt: Date.now(),
       });
+      res.json(updateCollege);
+    } catch ({ message: errMessage }) {
+      const message = errMessage ? errMessage : UNKNOWN_ERROR_OCCURED;
+      res.status(500).json(message);
     }
   } else {
-    res.json({
-      dbRes: "College Cannot be found",
-      isSuccess: false
-    });
+    res.status(500).json("College cannot be found");
   }
 });
 
@@ -178,8 +156,8 @@ router.delete("/:id", async (req, res) => {
     const getCollege = await College.find({
       _id: req.params.id,
       deletedAt: {
-        $exists: false
-      }
+        $exists: false,
+      },
     });
     if (getCollege.length > 0) {
       const deleteCollege = await College.findByIdAndUpdate(req.params.id, {
@@ -187,21 +165,13 @@ router.delete("/:id", async (req, res) => {
           deletedAt: Date.now(),
         },
       });
-      res.json({
-        dbRes: deleteCollege,
-        isSuccess: true
-      });
+      res.json(deleteCollege);
     } else {
-      res.json({
-        dbRes: "College is already deleted",
-        isSuccess: false
-      });
+      res.status(500).json("College is already deleted");
     }
-  } catch (error) {
-    res.json({
-      dbRes: error.message,
-      isSuccess: false
-    });
+  } catch ({ message: errMessage }) {
+    const message = errMessage ? errMssage : UNKNOWN_ERROR_OCCURED;
+    res.status(500).json(message);
   }
 });